# Enable link-time optimization