--- a/src/rust/engine/Cargo.toml
+++ b/src/rust/engine/Cargo.toml
@@ -10,6 +10,7 @@ build = "src/cffi_build.rs"
 debug = true
 # Optimise for the speed of our binary, rather than the speed of compilation.
 codegen-units = 1
+lto = true
 
 [lib]
 crate-type = ["cdylib"]