Before this change:
OSX:
 - Binary size: 2.9M
 - Time to compile and link (release): 348s
 - Time to list a large repo: 4m12

Linux:
 - Binary size: 51M
 - Time to compile and link (release): 556s
 - Time to list a large repo: 4m7

After this change:
OSX:
 - Binary size: 3M
 - Time to compile and link (release): 391s
 - Time to list a large repo: 4m8

Linux:
 - Binary size: 19M
 - Time to compile and link (release): 628s
 - Time to list a large repo: 4m7