@illicitonion 2018-03-06
Depends on https://github.com/pantsbuild/pants/pull/5555

It's not obvious that this is a PR we want to merge. It's effect on linux binary size is very significant, but it slows down linking a _lot_, particularly on Linux. I guess what I really want is a 3rd profile type: `dev`, `fast-to-build-and-run`, and `release`, but cargo only offers us `dev` and `release`.