@stuhood 2018-03-07
I don't think binary size matters enough to do this without some other performance improvement. Can revisit in the future?